fix: fix Mock test base
